factor out key down handler
